Added new Step to tweak background audio and some other sound effects

+++ utils/audiogeneration.js
@@ -4,14 +4,14 @@ import FfmpegCommand from 'fluent-ffmpeg';
 import fs from 'fs';
 
 
-  const generateCoolBoardRadioEffect = (inputFilePath) => {
+  const generateCoolBoardRadioEffect = (inputFilePath, notification, backgroundSound, backgroundVolume) => {
    return new Promise(function(resolve, reject) {
     const command = new FfmpegCommand();
     const command2 = new FfmpegCommand();
     //create audio file with voice and engine sound under.
     command
     .addInput(inputFilePath)
-    .addInput('./audio/utility/RB16BengineOnboard.mp3').seekInput(52)
+    .addInput(`./audio/utility/${backgroundSound}`).inputOption("-stream_loop -1").seekInput(42)
     //'amix=inputs=2:duration=first:dropout_transition=0'
     .complexFilter([{
       filter: 'volume',
@@ -21,7 +21,7 @@ import fs from 'fs';
     },
     {
       filter: 'volume',
-      options: ['0.1'],
+      options: [`${backgroundVolume}`],
       inputs: "1:0",
       outputs: "[s2]"
     },
@@ -35,27 +35,40 @@ import fs from 'fs';
     .on('end', function() {
       console.log('Amixed audio files together.');
 
-      //put noti sound in front of previous generated file.
-      command2.addInput('./audio/utility/f1radionotification.mp3').addInput('./audio/temp/overlayed.mp3').on('error', function(err) {
-        console.log('An error occurred: ' + err.message);
-      })
-      .on('end', function() {
-        console.log('Merged audio files successfully!');
+        if(notification == true)
+        {
+          //put noti sound in front of previous generated file.
+          command2.addInput('./audio/utility/f1radionotification.mp3').addInput('./audio/temp/overlayed.mp3').on('error', function(err) {
+            console.log('An error occurred: ' + err.message);
+          })
+          .on('end', function() {
+            console.log('Merged audio files successfully!');
 
-        fs.unlink('./audio/temp/overlayed.mp3', (err) => {
-          if (err) {
-            reject(new Error(err?.message || "Whoops!"));
-          } 
-          //file removed
-          resolve();
-        })
-      })
-      .mergeToFile('./audio/preview.mp3', './audio/temp');
-    });
+            fs.unlink('./audio/temp/overlayed.mp3', (err) => {
+              if (err) {
+                reject(new Error(err?.message || "Whoops!"));
+              } 
+              //file removed
+              resolve();
+            });
+          })
+          .mergeToFile('./audio/preview.mp3', './audio/temp');
+        }
+        else {
+          //move the overlayed audio file to audio and call it preview
+          fs.rename('./audio/temp/overlayed.mp3', './audio/preview.mp3', (err) => {
+            if (err) {
+              reject(new Error(err?.message || "Whoops!"));
+            } 
+            //file removed
+            resolve();
+          });
+        }
+      });
     });
   };
 
-  const generateCoolBoardRadioEffectWithDistortion = (inputFilePath) => {
+  const generateCoolBoardRadioEffectWithDistortion = (inputFilePath, notification, backgroundSound, backgroundVolume) => {
     return new Promise(function(resolve, reject) {
      const command = new FfmpegCommand();
      const command2 = new FfmpegCommand();
@@ -99,17 +112,17 @@ import fs from 'fs';
      .on('end', function() {
         command2
         .addInput('./audio/temp/distort.mp3')
-        .addInput('./audio/utility/RB16BengineOnboard.mp3').seekInput(52)
+        .addInput(`./audio/utility/${backgroundSound}`).seekInput(52)
         //'amix=inputs=2:duration=first:dropout_transition=0'
         .complexFilter([{
           filter: 'volume',
-          options: ['2.0'],
+          options: ['1.0'],
           inputs: "0:0",
           outputs: "[s1]"
         },
         {
           filter: 'volume',
-          options: ['0.1'],
+          options: [`${backgroundVolume}`],
           inputs: "1:0",
           outputs: "[s2]"
         },
@@ -123,14 +136,34 @@ import fs from 'fs';
         .on('end', function() {
           console.log('Amixed audio files together.');
     
-          //put noti sound in front of previous generated file.
-          command3.addInput('./audio/utility/f1radionotification.mp3').addInput('./audio/temp/overlayed.mp3').on('error', function(err) {
-            console.log('An error occurred: ' + err.message);
-          })
-          .on('end', function() {
-            console.log('Merged audio files successfully!');
-    
-            fs.unlink('./audio/temp/overlayed.mp3', (err) => {
+            if(notification)
+            {
+              //put noti sound in front of previous generated file.
+              command3.addInput('./audio/utility/f1radionotification.mp3').addInput('./audio/temp/overlayed.mp3').on('error', function(err) {
+                console.log('An error occurred: ' + err.message);
+              })
+              .on('end', function() {
+                console.log('Merged audio files successfully!');
+        
+                fs.unlink('./audio/temp/overlayed.mp3', (err) => {
+                  if (err) {
+                    reject(new Error(err?.message || "Whoops!"));
+                  } 
+                  fs.unlink('./audio/temp/distort.mp3', (err) => {
+                    if (err) {
+                      reject(new Error(err?.message || "Whoops!"));
+                    }      
+                    //files removed
+                    resolve();
+                  });
+                });
+
+              })
+              .mergeToFile('./audio/preview.mp3', './audio/temp');
+          }
+          else {
+            //move the overlayed audio file to audio and call it preview
+            fs.rename('./audio/temp/overlayed.mp3', './audio/preview.mp3', (err) => {
               if (err) {
                 reject(new Error(err?.message || "Whoops!"));
               } 
@@ -142,9 +175,7 @@ import fs from 'fs';
                 resolve();
               });
             });
-
-          })
-          .mergeToFile('./audio/speechWithCoolBoardRadioEffectDistorted.mp3', './audio/temp');
+          }
         });
         });
     });
